+++ lib/generators/rodauth/templates/config/initializers/sequel.rb
@@ -1,7 +1,7 @@
 require "sequel/core"
 
 # initialize Sequel and have it reuse Active Record's database connection
-<%- if RUBY_ENGINE == "jruby" -%>
+<% if RUBY_ENGINE == "jruby" -%>
 DB = Sequel.connect("jdbc:<%= sequel_adapter %>://", extensions: :activerecord_connection)
 <% else -%>
 DB = Sequel.<%= sequel_adapter %>(extensions: :activerecord_connection)