+++ test/generators/install_generator_test.rb
@@ -43,6 +43,23 @@ class InstallGeneratorTest < Rails::Generators::TestCase
     run_generator
 
     assert_file "app/lib/rodauth_app.rb", /class RodauthApp < Rodauth::Rails::App/
+    assert_file "app/lib/rodauth_app.rb", /configure do/
+    assert_file "app/lib/rodauth_app.rb", /:login, :logout, :remember,/
+    assert_file "app/lib/rodauth_app.rb", /flash_notice_key/
+    assert_file "app/lib/rodauth_app.rb", /Remember Feature/
+    assert_file "app/lib/rodauth_app.rb", /rodauth\.load_memory/
+  end
+
+  test "app with API only" do
+    Rails.application.config.api_only = true
+
+    run_generator
+
+    assert_file "app/lib/rodauth_app.rb", /configure json: :only do/
+    assert_file "app/lib/rodauth_app.rb", /:login, :logout, :jwt,/
+    assert_file "app/lib/rodauth_app.rb", /jwt_secret "[a-z0-9]{128}"/
+
+    Rails.application.config.api_only = false
   end
 
   test "controller" do